requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pinecone
requests
PyMuPDF
streamlit
openai
python-multipart